Change crypto_length prototype to return unsigned int, so it can be used directly with krb5_data
git-svn-id: svn://anonsvn.mit.edu/krb5/branches/mskrb-integ@21123 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/include/k5-int.h
@@ -622,7 +622,7 @@ struct krb5_aead_provider {
 				      const struct krb5_enc_provider *enc,
 				      const struct krb5_hash_provider *hash,
 				      krb5_cryptotype type,
-				      size_t *length);
+				      unsigned int *length);
     krb5_error_code (*encrypt_iov) (const struct krb5_aead_provider *aead,
 				    const struct krb5_enc_provider *enc,
 				    const struct krb5_hash_provider *hash,
